Moved generic http clients to separate module to reuse between stores when needed.
