+++ mlte/store/artifact/factory.py
@@ -6,7 +6,7 @@ Top-level functions for artifact store creation.
 
 from mlte.store.artifact.store import ArtifactStore
 from mlte.store.artifact.underlying.fs import LocalFileSystemStore
-from mlte.store.artifact.underlying.http import RemoteHttpStore
+from mlte.store.artifact.underlying.http import HttpArtifactStore
 from mlte.store.artifact.underlying.memory import InMemoryStore
 from mlte.store.artifact.underlying.rdbs.store import RelationalDBStore
 from mlte.store.base import StoreType, StoreURI
@@ -24,7 +24,7 @@ def create_store(uri: str) -> ArtifactStore:
     if parsed_uri.type == StoreType.LOCAL_FILESYSTEM:
         return LocalFileSystemStore(parsed_uri)
     if parsed_uri.type == StoreType.REMOTE_HTTP:
-        return RemoteHttpStore(parsed_uri)
+        return HttpArtifactStore(parsed_uri)
     if parsed_uri.type == StoreType.RELATIONAL_DB:
         return RelationalDBStore(parsed_uri)
     else:
